requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
django-phone-verify==2.0.1
django-phonenumber-field==7.0.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
nexmo==2.5.2
oauthlib==3.2.1
phonenumbers==8.12.55
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyotp==2.7.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.14.0
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
